+++ Assets/Machinator/Scripts/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    void ApplyDamage(float amount);
+    void ApplyDamage(int amount);
 }
